Message Scrolling across the screen
In this practice a code was created, which if a button is pressed on the transmitter, a message is sent, which the receiver will interpret and put an initial name on the screen, and scroll it across the entire screen until it exits. Then the other name will appear on the other side and will scroll across the screen until it exits, and then repeat the program again.