Message displayed after counter of 10 seconds
In this practice we created a code, which press a button on the transmitter, a message is sent; which the receiver will interpret and put an initial name on the screen, and side a counter, that after 10 seconds clear the screen and the other name (or the same) will appear under the original. And when you press again the button, the program repeat the action.